store/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.9.4
django-cors-headers>=3.8.0